Add implementation of entities standarts (doesn't work yet)
There are some huge changes to data structures.
Now we have Primitive, and void pointer to primitive's data.
We gonna provide new structure for each entity which we gonna implement.
And not only structure, but also a func to fill it.
Entities standarts are in structures in entities_structs.h. It's very nice idea to have all the standarts in such way. Gonna polish this.